backend/pyproject.toml
[tool.poetry]
name = "fastapi-ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
langchain = "^0.3.13"
qdrant-client = "^1.12.2"
python-decouple = "^3.8"
langchain-community = "^0.3.13"
openai = "^1.58.1"
beautifulsoup4 = "^4.12.3"
tiktoken = "^0.8.0"
langchain-openai = "^0.2.14"
langchain-qdrant = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
